Update README.md and .env.example
The .env file doesn't need to be sourced, if you run it only in docker.
Docker compose directly reads from it.

Our gitlab ip is afaik pretty stable, so it should be fine to just add
it.

getent is more commonly installed compared to dig :)

I added a section on how to enable the compose plugin for docker.