Issue #135: add CORS middleware for the backend APIs
For production and dev, we will have requests coming from third-party
clients that will need CORS headers in order to properly process
responses from the API:

* Authorization API: from any source
* Management API: Management UI

This commit adds CORS middleware to the authz and management APIs, if
they are configured to do so.

Additionally, during development work for the Management UI, we can use
a proxy to develop against an external server (although this is not
secure for production).