Issue #185: fix traceback in logs when database isn't ready
The `UndefinedTableError` found in the logs when first installing the
Management API is harmless. It happens because alembic isn't finished
running its migrations when the bundle server is trying to build the
bundles for the first time. Since the bundle server keeps retrying every
few seconds, it will eventually build the bundle without issue.

However, we do log the traceback chain from the db, which clutters up the
logs with something that looks serious, but really isn't. Since we
expect this to happen when first installing the app, it makes sense to
turn it into a warning (so we can see if we have too many of these), but
not print the full traceback chain.