Issue #234: remove references to Vagrant
Per discussion with Ole Schweigert, we have decided not to support
Vagrant as a development option because few team members were interested
in its use.

In order to meet the acceptance criteria for this issue, we need the
development setup documents to be correct. For this reason, we have
removed all references to Vagrant from the project.

For those team members who wish to keep using Vagrant, please add the
relevant files to your personal `.gitignore`.