Issue #237: configure Keycloak token correctly for 25.0.4
In UCS we have upgraded Keycloak to 25.0.4. This caused some breaking changes
in the Guardian, because the "sub" key is no longer present in Keycloak tokens.
This was a breaking change between Keycloak 24 and 25.

This commit updates Guardian to expect the "username" key instead. This should
be backwards-compatible, in that the "username" mapper gets added even if the
"sub" key is still present.