fix: add basic scope to guardian-ui client
With Keycloak 25+ [source](https://www.keycloak.org/docs/latest/release_notes/#lightweight-access-token-to-be-even-more-lightweight),
there is a new client scope `basic` which add claims `sub` and `auth_time`.
This scope should be added to all clients automatically, but since we explicitly exclude
it in our provisioning, it is not added.

Issue univention/components/authorization-engine/guardian#237